requirements.txt
flask
google-generativeai
speechrecognition
pyttsx3
requests
beautifulsoup4
googlesearch-python
